+++ index.html
@@ -36,29 +36,9 @@
             <li class="nav-item">
               <a class="nav-link" href="#">Profile</a>
             </li>
-            <li class="nav-item dropdown">
-              <a
-                class="nav-link dropdown-toggle"
-                href="#"
-                role="button"
-                data-bs-toggle="dropdown"
-                aria-expanded="false"
-              >
-                Dropdown
-              </a>
-              <ul class="dropdown-menu">
-                <li><a class="dropdown-item" href="#">Action</a></li>
-                <li><a class="dropdown-item" href="#">Another action</a></li>
-                <li>
-                  <hr class="dropdown-divider" />
-                </li>
-                <li>
-                  <a class="dropdown-item" href="#">Something else here</a>
-                </li>
-              </ul>
-            </li>
+
             <li class="nav-item">
-              <a class="nav-link disabled">Disabled</a>
+              <a class="nav-link">foo</a>
             </li>
           </ul>
           <form class="d-flex" role="search">
